fix: Access list saving and DB migration improvements
- Fix access_list_id not saving when applied to proxy hosts
- Add access_list_id to Create/Update repository functions
- Fix rule_id integer overflow (INTEGER -> BIGINT) for ModSecurity rules
- Support TimescaleDB: decompress chunks before ALTER, recompress after
- Change challenge_tokens FK to ON DELETE SET NULL
- Change Go code RuleID type from int to int64 (model, repository, handler, service)
- Add BlockReasonCloudProviderBlock constant
- Bump version to 1.3.9

+++ api/internal/service/waf_auto_ban.go
@@ -113,7 +113,7 @@ func (s *WAFAutoBanService) refreshSettings(ctx context.Context) {
 }
 
 // RecordWAFEvent records a WAF event for an IP and checks if it should be banned
-func (s *WAFAutoBanService) RecordWAFEvent(ctx context.Context, clientIP string, host string, ruleID int, ruleMessage string) {
+func (s *WAFAutoBanService) RecordWAFEvent(ctx context.Context, clientIP string, host string, ruleID int64, ruleMessage string) {
 	s.settingsMu.RLock()
 	enabled := s.enabled
 	threshold := s.threshold
